# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SILAKAP API",
    "version": "1.0.0",
    "description": "Dokumentasi API Sistem Informasi Layanan Administrasi Kepegawaian."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API prefix"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/integrasi/import/asn": {
      "post": {
        "tags": [
          "Integrasi"
        ],
        "summary": "Import data ASN dari Excel",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Import ASN diproses"
          }
        }
      }
    },
...
    "/integrasi/log/{id}/errors": {
      "get": {
        "tags": [
          "Integrasi"
        ],
        "summary": "Export error import ke Excel",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File Excel error import"
          }
        }
      }
    },
...
  }
}